complete category

+++ admin/category.php
@@ -1,4 +1,8 @@
-<?php include "header.php"; ?>
+<?php include "header.php"; 
+    if ($_SESSION['role'] == '0') {
+        header('location: post.php');
+    }
+?>
 <div id="admin-content">
     <div class="container">
         <div class="row">
@@ -9,6 +13,18 @@
                 <a class="add-new" href="add-category.php">add category</a>
             </div>
             <div class="col-md-12">
+
+            <?php
+                  include "config.php";
+
+                 
+                  $query = "SELECT * FROM category ORDER BY category_id DESC ";
+                  $result = mysqli_query($connection, $query) or die('Query Field');
+                  $count = mysqli_num_rows($result);
+
+                  if ($count > 0) {
+                    
+                ?>
                 <table class="content-table">
                     <thead>
                         <th>S.No.</th>
@@ -18,50 +34,26 @@
                         <th>Delete</th>
                     </thead>
                     <tbody>
+                    <?php 
+                        while ($row = mysqli_fetch_assoc($result)) {
+                          
+                        $serial_number = 1;
+                      ?>
                         <tr>
-                            <td class='id'>1</td>
-                            <td>Html</td>
-                            <td>5</td>
-                            <td class='edit'><a href='update-category.php'><i class='fa fa-edit'></i></a></td>
-                            <td class='delete'><a href='delete-category.php'><i class='fa fa-trash-o'></i></a></td>
-                        </tr>
-                        <tr>
-                            <td class='id'>2</td>
-                            <td>Css</td>
-                            <td>15</td>
-                            <td class='edit'><a href='update-category.php'><i class='fa fa-edit'></i></a></td>
-                            <td class='delete'><a href='delete-category.php'><i class='fa fa-trash-o'></i></a></td>
-                        </tr>
-                        <tr>
-                            <td class='id'>3</td>
-                            <td>Java</td>
-                            <td>8</td>
-                            <td class='edit'><a href='update-category.php'><i class='fa fa-edit'></i></a></td>
-                            <td class='delete'><a href='delete-category.php'><i class='fa fa-trash-o'></i></a></td>
-                        </tr>
-                        <tr>
-                            <td class='id'>4</td>
-                            <td>Php</td>
-                            <td>11</td>
-                            <td class='edit'><a href='update-category.php'><i class='fa fa-edit'></i></a></td>
-                            <td class='delete'><a href='delete-category.php'><i class='fa fa-trash-o'></i></a></td>
-                        </tr>
-                        <tr>
-                            <td class='id'>5</td>
-                            <td>Python</td>
-                            <td>13</td>
-                            <td class='edit'><a href='update-category.php'><i class='fa fa-edit'></i></a></td>
-                            <td class='delete'><a href='delete-category.php'><i class='fa fa-trash-o'></i></a></td>
-                        </tr>
-                        <tr>
-                            <td class='id'>6</td>
-                            <td>Scss</td>
-                            <td>3</td>
-                            <td class='edit'><a href='update-category.php'><i class='fa fa-edit'></i></a></td>
-                            <td class='delete'><a href='delete-category.php'><i class='fa fa-trash-o'></i></a></td>
+                            <td class='id'><?php echo $serial_number++ ?></td>
+                            <td><?php echo $row['category_name'] ?></td>
+                            <td><?php echo $row['category_id'] ?></td>
+                            <td class='edit'><a href='update-category.php?id=<?php echo $row['category_id'] ?>'><i class='fa fa-edit'></i></a></td>
+                            <td class='delete'><a onclick="return confirm('Are You Sure?')" href='delete-category.php?id=<?php echo $row['category_id'] ?>'><i class='fa fa-trash-o'></i></a></td>
                         </tr>
                     </tbody>
+
+                    <?php } ?>
+                    
                 </table>
+
+                <?php } ?>
+
                 <ul class='pagination admin-pagination'>
                     <li class="active"><a>1</a></li>
                     <li><a>2</a></li>
